feat: Kotlin Microservice 개발 예제3 작성

+++ yolo-api/src/main/kotlin/com/yolo/jean/customer/api/CustomerApi.kt
@@ -3,6 +3,8 @@ package com.yolo.jean.customer.api
 import com.yolo.jean.customer.domain.Customer
 import com.yolo.jean.customer.service.CustomerService
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 
 
@@ -13,24 +15,43 @@ class CustomerApi {
     lateinit var customerService: CustomerService
 
     @GetMapping(value = ["/customer/{id}"])
-    fun getCustomer(@PathVariable id: Int) = customerService.getCustomer(id)
+    fun getCustomer(@PathVariable id: Int): ResponseEntity<Customer?> {
+        val customer = customerService.getCustomer(id)
+        val status = if (customer == null) HttpStatus.NOT_FOUND else HttpStatus.OK
+        return ResponseEntity(customer, status)
+    }
 
     @PostMapping(value = ["/customer/"])
-    fun createCustomer(@RequestBody customer: Customer) {
+    fun createCustomer(@RequestBody customer: Customer): ResponseEntity<Unit?> {
         customerService.createCustomer(customer)
+        return ResponseEntity(null, HttpStatus.CREATED)
     }
 
     @PutMapping(value = ["/customer/{id}"])
-    fun updateCustomer(@PathVariable id:Int, @RequestBody customer: Customer) {
-        customerService.updateCustomer(id, customer)
+    fun updateCustomer(@PathVariable id: Int, @RequestBody customer: Customer): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+
+        if (customerService.getCustomer(id) != null) {
+            customerService.updateCustomer(id, customer)
+            status = HttpStatus.ACCEPTED
+        }
+        return ResponseEntity(Unit, status)
     }
 
     @DeleteMapping(value = ["/customer/{id}"])
-    fun deleteCustomer(@PathVariable id:Int) {
-        customerService.deleteCustomer(id)
+    fun deleteCustomer(@PathVariable id: Int): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+
+        if (customerService.getCustomer(id) != null) {
+            customerService.deleteCustomer(id)
+            status = HttpStatus.OK
+        }
+        return ResponseEntity(Unit, status)
     }
 
     @GetMapping(value = ["/customers"])
-    fun getCustomers(@RequestParam(required = false, defaultValue = "")
-    nameFilter: String) = customerService.searchCustomers(nameFilter)
+    fun getCustomers(
+        @RequestParam(required = false, defaultValue = "")
+        nameFilter: String
+    ) = customerService.searchCustomers(nameFilter)
 }
